Allow virtual machines to use files in project directory as disk images
This is useful for Cisco VMs that will load a boot configuration if it
appears on an ISO filesystem attached as a virtual CD-ROM, and boot
configurations are logically part of the project, not global.

Part of this change is that the server no longer tries to relativize
the paths before returning them in JSON objects.